Tela LoadAuth feita, Cor dor backgournd do app configurada, cobranca_25-01-2020.apk
