+++ App.js
@@ -1,9 +1,8 @@
 import React, { useState, useEffect } from "react";
-import { View, StyleSheet, Text } from "react-native";
-import { Spinner } from "native-base";
 import * as Font from "expo-font";
 
 import Routes from "./src/routes";
+import Loading from "./src/components/Loading";
 
 export default function App() {
   const [loading, setLoading] = useState(true);
@@ -20,29 +19,8 @@ export default function App() {
   });
 
   if (loading) {
-    return (
-      <View style={styles.container}>
-        <Spinner color="#F3F3F3" />
-        <Text style={styles.textLoading}>Carregando...</Text>
-      </View>
-    );
+    return <Loading />;
   } else {
     return <Routes />;
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#444444",
-    alignItems: "center",
-    justifyContent: "center"
-  },
-  textLoading: {
-    paddingTop: 15,
-    fontSize: 20,
-    color: "#F3F3F3",
-    alignItems: "center",
-    justifyContent: "center"
-  }
-});
