+++ app.json
@@ -4,14 +4,14 @@
     "slug": "cobranca",
     "privacy": "public",
     "sdkVersion": "36.0.0",
-    "platforms": ["ios", "android", "web"],
+    "platforms": ["ios", "android"],
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "splash": {
       "image": "./assets/splash.png",
       "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
+      "backgroundColor": "#444"
     },
     "updates": {
       "fallbackToCacheTimeout": 0
@@ -22,6 +22,8 @@
     },
     "android": {
       "package": "com.alexsouzasilvax.cobranca"
-    }
+    },
+    "githubUrl": "https://github.com/alexsouzasilvax/app-cobranca",
+    "primaryColor": "#444"
   }
 }
